requirements.txt
flake8
pytest
merakicommons
datapipelines
pycurl
Pillow